Initial commit: CV monitor foundation
Created a quick foundation of the code for the LifeLine CV monitoring system. We still need to create a neat frontend - the outputs right now are fake demo since the model isn't trained on any real data so it doesn't work properly. Also we have to follow HIPAA rules so we need to encrypt all the data/images it captures and maybe implement auto-delete features. Other features are still needed but this gets us started with the basic structure.